9/5/19

+++ src/store/reducers/auth.js
@@ -1,24 +1,26 @@
 import { CLIENT_SET, CLIENT_UNSET } from "../constants/auth";
 import jwtDecode from "jwt-decode";
-let data = localStorage.getItem("token")
+let data = typeof window !== 'undefined' && localStorage.getItem("token")
   ? jwtDecode(localStorage.getItem("token"))
   : {};
 const initialSate = {
   user: data,
-  token: localStorage.getItem("token")
+  token:  typeof window !== 'undefined' ? localStorage.getItem("token"): null
 };
 console.log("dataaa=====", data);
 const reducer = function clientReducer(state = initialSate, action) {
   switch (action.type) {
     case CLIENT_SET:
-      console.log('sagaaaaL ',action)
+      console.log('sagaaaaL ', action)
       return {
         user: jwtDecode(action.token),
         token: action.token
       };
 
     case CLIENT_UNSET:
-      localStorage.clear()
+      if (typeof window !== 'undefined') {
+        localStorage.clear()
+      }
       return {
         user: null,
         token: null
